Continue graph decomposition by extracting entity query surface (#295)
